requirements.txt
aiomysql==0.1.1
aioredis==2.0.1
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
autopep8==1.7.0
bcrypt==4.0.0
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.81.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
Mako==1.2.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.9.1
pydantic==1.9.2
PyMySQL==1.0.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
toml==0.10.2
typing-extensions==4.3.0
uvicorn==0.18.3
